refactor: fix test SUT names + move group to proper file
