+++ internal/integrations/v4/runner/runner_load.go
@@ -6,33 +6,12 @@ import (
 	"github.com/newrelic/infrastructure-agent/internal/integrations/v4/integration"
 	"github.com/newrelic/infrastructure-agent/pkg/databind/pkg/databind"
 	config2 "github.com/newrelic/infrastructure-agent/pkg/integrations/v4/config"
-	"github.com/newrelic/infrastructure-agent/pkg/integrations/v4/emitter"
 )
 
 // Loader provides a basic, incomplete Group instance to be configured by the
 // NewGroup function. The DefinitionsRepo instance is only required to load
 // v3 legacy integrations from an external definitions' folder.
 type Loader func(dr integration.InstancesLookup, passthroughEnv []string, cfgPath string) (Group, FeaturesCache, error)
-
-// NewGroup configures a Group instance that is provided by the passed Loader
-// cfgPath is used for caching to be consumed by cmd-channel FF enabler.
-func NewGroup(
-	loader Loader,
-	dr integration.InstancesLookup,
-	passthroughEnv []string,
-	emitter emitter.Emitter,
-	cfgPath string,
-) (g Group, c FeaturesCache, err error) {
-
-	g, c, err = loader(dr, passthroughEnv, cfgPath)
-	if err != nil {
-		return
-	}
-
-	g.emitter = emitter
-
-	return
-}
 
 // LoadFrom returns a partial Group that holds the configuration from the provided configuration.
 // Optionally agent and OHI "features" can be provided to be able to load disabled OHIs.
